+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/throughputControl/ThroughputControlTests.java
@@ -89,7 +89,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         };
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputLocalControl_requestOptions(OperationType operationType) {
         this.ensureContainer();
         // The create document in this test usually takes around 6.29RU, pick a RU here relatively close, so to test throttled scenario
@@ -117,7 +117,7 @@ public class ThroughputControlTests extends TestSuiteBase {
             BridgeInternal.getContextClient(client).getConnectionPolicy().getConnectionMode());
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputLocalControl_default(OperationType operationType) {
         this.ensureContainer();
 
@@ -155,7 +155,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, timeOut = TIMEOUT)
     public void throughputLocalControlWithThroughputQuery() {
         // Will need to use a new client here to make sure the throughput query mono will be passed down to throughputContainerController
         CosmosAsyncClient cosmosAsyncClient = null;
@@ -205,7 +205,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputLocalControlPriorityLevel(OperationType operationType) {
         ThroughputControlGroupConfig groupConfig =
             new ThroughputControlGroupConfigBuilder()
@@ -228,7 +228,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         assertThat(createItemResponse.getStatusCode()).isEqualTo(201);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputGlobalControl(OperationType operationType) {
         this.ensureContainer();
         String controlContainerId = "tcc" + UUID.randomUUID();
@@ -272,7 +272,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, timeOut = TIMEOUT)
     public void throughputGlobalControlWithThroughputQuery() {
         this.ensureContainer();
         // Will need to use a new client here to make sure the throughput query mono will be passed down to throughputContainerController
@@ -330,7 +330,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputGlobalControlCanUpdateConfig(OperationType operationType) {
         this.ensureContainer();
         String controlContainerId = "tcc" + UUID.randomUUID();
@@ -398,7 +398,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "operationTypeProvider", timeOut = TIMEOUT)
     public void throughputLocalControlForContainerCreateDeleteWithSameName(OperationType operationType) throws InterruptedException {
         this.ensureContainer();
         ConnectionMode connectionMode = BridgeInternal.getContextClient(client).getConnectionPolicy().getConnectionMode();
@@ -466,7 +466,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, timeOut = TIMEOUT)
     public void throughputLocalControl_createItem() throws InterruptedException {
         this.ensureContainer();
         // The create document in this test usually takes around 6.29RU, pick a RU here relatively close, so to test throttled scenario
@@ -502,7 +502,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         validateItemSuccess(container.createItem(itemGetThrottled), successValidator);
     }
 
-    @Test(groups = {"emulator"}, dataProvider = "allowRequestToContinueOnInitErrorProvider", timeOut = TIMEOUT)
+    @Test(groups = {"long-emulator"}, dataProvider = "allowRequestToContinueOnInitErrorProvider", timeOut = TIMEOUT)
     public void throughputControlContinueOnInitError(boolean continueOnInitError) {
         this.ensureContainer();
         // Purposely not creating the throughput control container so to test allowRequestContinueOnInitError
@@ -542,7 +542,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, timeOut = TIMEOUT * 4)
+    @Test(groups = {"long-emulator"}, timeOut = TIMEOUT * 4)
     public void throughputGlobalControlMultipleClients() throws InterruptedException {
         this.ensureContainer();
         List<CosmosAsyncClient> cosmosAsyncClients = new ArrayList<>();
@@ -596,7 +596,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @Test(groups = {"emulator"}, timeOut = TIMEOUT * 4)
+    @Test(groups = {"long-emulator"}, timeOut = TIMEOUT * 4)
     public void enableSameGroupMultipleTimes() {
         this.ensureContainer();
 
@@ -670,7 +670,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @BeforeClass(groups = { "emulator" }, timeOut = 4 * SETUP_TIMEOUT)
+    @BeforeClass(groups = { "long-emulator" }, timeOut = 4 * SETUP_TIMEOUT)
     public void before_ThroughputBudgetControllerTest() {
         this.ensureContainer();
     }
@@ -697,7 +697,7 @@ public class ThroughputControlTests extends TestSuiteBase {
         }
     }
 
-    @AfterClass(groups = {"emulator"}, timeOut = TIMEOUT, alwaysRun = true)
+    @AfterClass(groups = {"long-emulator"}, timeOut = TIMEOUT, alwaysRun = true)
     public void after_ThroughputBudgetControllerTest() {
         safeClose(this.client);
     }