+++ sdk/netapp/azure-resourcemanager-netapp/src/main/java/com/azure/resourcemanager/netapp/implementation/AccountsClientImpl.java
@@ -2540,6 +2540,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a subscription
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.
@@ -2568,6 +2570,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a subscription
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.
@@ -2595,6 +2599,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a subscription
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.
@@ -2623,6 +2629,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a resource group
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.
@@ -2649,6 +2657,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a resource group
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.
@@ -2676,6 +2686,8 @@ public final class AccountsClientImpl implements AccountsClient {
     }
 
     /**
+     * Describe all NetApp Accounts in a resource group
+     * 
      * Get the next page of items.
      * 
      * @param nextLink The URL to get the next list of items.