+++ sdk/spring/spring-cloud-azure-stream-binder-eventhubs/src/test/java/com/azure/spring/cloud/stream/binder/eventhubs/implementation/config/EventHubsBinderConfigurationTests.java
@@ -93,6 +93,9 @@ public class EventHubsBinderConfigurationTests {
             });
     }
 
+    // conniey: Remove warning suppression when azure-messaging-eventhubs is updated to 5.21.0.
+    // https://github.com/Azure/azure-sdk-for-java/issues/46359
+    @SuppressWarnings("deprecation")
     @Test
     void testExtendedBindingPropertiesShouldBind() {
         String producerConnectionString = String.format(CONNECTION_STRING_FORMAT, "fake-producer-namespace");