+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -1,12 +1,22 @@
 # Release History
 
-## 1.21.0-beta.2 (Unreleased)
+## 1.22.0-beta.1 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
 ### Bugs Fixed
+
+### Other Changes
+
+## 1.21.0 (2025-08-11)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-messaging-eventhubs` from `5.20.5` to version `5.21.0`.
 
 ## 1.20.10 (2025-07-29)
 
@@ -15,7 +25,6 @@
 #### Dependency Updates
 
 - Upgraded `azure-messaging-eventhubs` from `5.20.5` to version `5.21.0-beta.2`.
-
 
 ## 1.20.9 (2025-07-24)