+++ sdk/resourcemanager/azure-resourcemanager-test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
-## 2.0.0-beta.1 (Unreleased)
+## 2.0.0-beta.2 (Unreleased)
 
-- TODO
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 2.0.0-beta.1 (2025-08-12)
+
+### Features Added
+
+- Initial release of the `azure-resourcemanager-test` library.